Blockquote does not contain q element
<q> element is for inline quotes. If it’s inside a block quote, then
semantically it’s read as a being a quote within a quote, which
unlikely to be the case very often. I’ve therefore replaced it (as well
as its related CSS) with a <p> tag instead).

As ever, apologies if I’ve got the wrong end of the stick though…